+++ packages/cli/src/commands/index.ts
@@ -1,4 +1,5 @@
 import { agentCommand } from './agent/command';
+import { agentRunsCommand } from './agent-runs/command';
 import { activityCommand } from './activity/command';
 import { aiGatewayCommand } from './ai-gateway/command';
 import { alertsCommand } from './alerts/command';
@@ -58,6 +59,7 @@ import { telemetryCommand } from './telemetry/command';
 import { tracesCommand } from './traces/command';
 import { upgradeCommand } from './upgrade/command';
 import { usageCommand } from './usage/command';
+import { vcrCommand } from './vcr/command';
 import { whoamiCommand } from './whoami/command';
 import { blobCommand } from './blob/command';
 import { webhooksCommand } from './webhooks/command';
@@ -66,6 +68,7 @@ import output from '../output-manager';
 
 const commandsStructs = [
   agentCommand,
+  agentRunsCommand,
   aiGatewayCommand,
   alertsCommand,
   aliasCommand,
@@ -124,6 +127,7 @@ const commandsStructs = [
   upgradeCommand,
   webhooksCommand,
   usageCommand,
+  vcrCommand,
   whoamiCommand,
   // added because we don't have a full help command
   { name: 'help', aliases: [] },