+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @vercel/cervel
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies [cbf22bf]
+  - @vercel/backends@0.8.21
+
+## 0.1.28
+
+### Patch Changes
+
+- @vercel/backends@0.8.20
+
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [62a884e]
+  - @vercel/backends@0.8.19
+
 ## 0.1.26
 
 ### Patch Changes